Вносим правки
