+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/service/AggregatorService.java
@@ -1,0 +1,9 @@
+package ru.yandex.practicum.service;
+
+import org.apache.avro.specific.SpecificRecordBase;
+import org.apache.kafka.clients.producer.Producer;
+import ru.yandex.practicum.kafka.telemetry.event.SensorEventAvro;
+
+public interface AggregatorService {
+    void sendSnapshot(Producer<String, SpecificRecordBase> producer, SensorEventAvro sensorEvent);
+}
