+++ telemetry/collector/src/main/java/ru/yandex/practicum/CollectorApp.java
@@ -2,14 +2,12 @@ package ru.yandex.practicum;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ConfigurableApplicationContext;
-import ru.yandex.practicum.aggregator.AggregationStarter;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class CollectorApp {
     public static void main(String[] args) {
-        ConfigurableApplicationContext context = SpringApplication.run(CollectorApp.class, args);
-        AggregationStarter aggregator = context.getBean(AggregationStarter.class);
-        aggregator.start();
+        SpringApplication.run(CollectorApp.class, args);
     }
 }
