+++ telemetry/serialization/avro-schemas/src/main/java/deserializer/SensorAvroDeserializer.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.kafka_client;
+package deserializer;
 
 import ru.yandex.practicum.kafka.telemetry.event.SensorEventAvro;
